Ledger

+++ e-permit/src/main/java/epermit/ledger/ledgerevents/permitrevoked/PermitRevokedLedgerEventHandler.java
@@ -1,7 +1,9 @@
 package epermit.ledger.ledgerevents.permitrevoked;
 
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 import epermit.ledger.entities.LedgerPermit;
+import epermit.ledger.ledgerevents.LedgerEventHandleResult;
 import epermit.ledger.ledgerevents.LedgerEventHandler;
 import epermit.ledger.repositories.LedgerPermitRepository;
 import lombok.RequiredArgsConstructor;
@@ -13,10 +15,21 @@ import lombok.extern.slf4j.Slf4j;
 public class PermitRevokedLedgerEventHandler implements LedgerEventHandler {
     private final LedgerPermitRepository permitRepository;
 
-    public void handle(Object e) {
+    public LedgerEventHandleResult handle(Object e) {
         log.info("PermitRevokedEventHandler started with {}", e);
         PermitRevokedLedgerEvent event = (PermitRevokedLedgerEvent) e;
-        LedgerPermit permit = permitRepository.findOneByPermitId(event.getPermitId()).get();
+        Optional<LedgerPermit> permitR = permitRepository.findOneByPermitId(event.getPermitId());
+        if (!permitR.isPresent()) {
+            log.info("PermitRevokedEventValidator result is INVALID_PERMITID");
+            return LedgerEventHandleResult.fail("INVALID_PERMITID");
+        }
+        LedgerPermit permit = permitR.get();
+        if (!(permit.getIssuedFor().equals(event.getIssuedFor())
+                && permit.getIssuer().equals(event.getIssuer()))) {
+            log.info("PermitRevokedEventValidator result is PERMIT_EVENT_MISMATCH");
+            return LedgerEventHandleResult.fail("PERMIT_EVENT_MISMATCH");
+        }
         permitRepository.delete(permit);
+        return LedgerEventHandleResult.success();
     }
 }
